Checkin even stale connections from reap
Previously when we reaped connections, we checked if they were inactive
(connected and responding to ping in most adapters) and if so we would
remove the connection instead of checking it back in.

Removing the connection would not wake a thread waiting on the queue for
a connection. This commit changes that to always re-check in the
connection.